Fix compilation missing override warnings and function declarations (#256)
This commit changes the signature of the log function in the DIakopticsSolver files to match the current ones used in the other solvers, that have now two arguments.

On the other side, the overrides on the SynchronGenerator files of the function clone had a different signature compared to all the rest of the classes (and did not override the Base class virtual function.

Signed-off-by: Leonardo Carreras <leonardo.carreras@eonerc.rwth-aachen.de>